requirements.txt
boto3
awscli
pypdf
langchain
langchain_community
streamlit
faiss-cpu

